# services/admin/openapi.yml
openapi: 3.0.0
info:
  description: 'The Admin service allows to manage API keys and unpaywall data.'
  version: 1.0.0
  title: Admin service
  contact:
    email: ezteam@couperin.org
    name: ezTeam
  license:
    name: CeCILL 2.1
    url: http://www.cecill.info/licences/Licence_CeCILL_V2.1-en.html
servers:
  - url: https://unpaywall.inist.fr/api/admin
...
paths:
...
  /job/download/insert/changefile/period:
    post:
      tags:
        - job
      operationId: post-update-job-period
      summary: Start process download and insert changefiles from unpaywall.
      description: Download and insert changefiles from unpaywall.org.
      security:
        - x-api-key: []
      requestBody:
        description: config for process
        content:
          application/json:
            schema:
              type: object
              required:
                - userName
              properties:
                index:
                  type: string
                interval:
                  type: string
                startDate:
                  type: string
                endDate:
                  type: string
                ignoreError:
                  type: boolean
                cleanFile:
                  type: boolean
      responses:
        '202':
          $ref: '#/components/responses/Accepted'
        '400':
          description: startDate cannot be in the futur or endDate cannot be lower than startDate.
        '401':
          $ref: '#/components/responses/Not-authorized'
        '409':
          $ref: '#/components/responses/Conflict'
  /job/insert/changefile/${filename}:
    post:
      tags:
        - job
      operationId: post-update-job-changefile-$-filename
      summary: Start process insert changesfiles downloaded on ezunpaywall.
      description: Insert changefiles already downloaded on ezunpaywall.
      security:
        - x-api-key: []
      parameters:
        - in: path
          name: filename
          description: filename
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                index:
                  type: string
                offset:
                  type: integer
                limit:
                  type: integer
                ignoreError:
                  type: boolean
                cleanFile:
                  type: boolean
        description: config for process
      responses:
        '202':
          $ref: '#/components/responses/Accepted'
        '400':
          description: startDate cannot be in the futur or endDate cannot be lower than startDate.
          content:
            application/json:
              examples:
                response:
                  value:
                    message: startDate cannot be in the futur or endDate cannot be lower than startDate.
        '401':
          $ref: '#/components/responses/Not-authorized'
        '404':
          $ref: '#/components/responses/File-not-found'
        '409':
          $ref: '#/components/responses/Conflict'
  /job/download/insert/history/period:
    post:
      tags:
        - job
      operationId: post-update-job-history
      summary: Start process download and insert changefiles from unpaywall and feed history.
      description: Download and insert changefiles from unpaywall.org and feed history.
      security:
        - x-api-key: []
      requestBody:
        description: config for process
        content:
          application/json:
            schema:
              type: object
              required:
                - userName
              properties:
                index:
                  type: string
                interval:
                  type: string
                startDate:
                  type: string
                endDate:
                  type: string
      responses:
        '202':
          $ref: '#/components/responses/Accepted'
        '400':
          description: startDate cannot be in the futur or endDate cannot be lower than startDate.
        '401':
          $ref: '#/components/responses/Not-authorized'
        '409':
          $ref: '#/components/responses/Conflict'
...
components:
  responses:
    Accepted:
      description: Accepted
      headers: {}
    Not-authorized:
      description: Not authorized
      headers: {}
      content:
        application/json:
          examples:
            response:
              value:
                message: Not authorized
    Conflict:
      description: Conflict
      headers: {}
      content:
        application/json:
          examples:
            response:
              value:
                message: Update in progress.
    File-not-found:
      description: File not found
      headers: {}
      content:
        application/json:
          examples:
            response:
              value:
                message: File not found.
...
  securitySchemes:
    x-api-key:
      type: apiKey
      in: header
      name: API Key
...